refactor: make WindowId an isize type
To reduce the number of casts

+++ native/desktop-macos/src/macos/window_api.rs
@@ -13,7 +13,7 @@ use super::{
     window::{NSWindowExts, Window},
 };
 
-pub type WindowId = i64;
+pub type WindowId = isize;
 
 pub type WindowPtr<'a> = RustAllocatedRawPtr<'a, std::ffi::c_void>;
